refactor: ContextWrapper -> ContextWrapperTest
